♻️ Extract composer auto-resize for Super Plan
Move autoResize into composer-auto-resize.ts so Super Plan can grow with
typed lines without importing the Code send graph. The JS fallback now
honors CSS min/max height, and Super Plan unbinds listeners on destroy.
